backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.3
tree-sitter-typescript==0.21.2
networkx==3.3
groq==0.9.0
httpx==0.27.2
python-dotenv==1.0.1
chardet==5.2.0
